Subscribe to node log stream on creating Breez foreground service
Unsubscribe from node log stream when shutting down Breez foreground service
